requirements.txt
Adafruit-LED-Backpack
Pillow
smbus
datetime
